+++ frontend/src/app/editor/page.tsx
@@ -67,7 +67,11 @@ export default function EditorPage() {
     canRedo,
 
     updateLayer,
+    addLayer,
+
     editSelectedLayers,
+    deleteSelectedLayers,
+    duplicateSelectedLayers,
   } = useLayersState();
 
   useEffect(() => {
@@ -129,7 +133,12 @@ export default function EditorPage() {
           layers,
           setVirtualLayers,
           updateLayer,
+          addLayer,
+
           editSelectedLayers,
+          deleteSelectedLayers,
+          duplicateSelectedLayers,
+
           commitVirtualLayers,
 
           getCanvasPointerPosition,
